requirements.txt
google-cloud-firestore
trino
pandas